% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddDominanceDataSheet.r
\name{AddDominanceDataSheet}
\alias{AddDominanceDataSheet}
\title{Function AddDominanceDataSheet}
\usage{
AddDominanceDataSheet(
  workbook,
  filename,
  sheetname,
  DATASheet,
  XLSX,
  overwrite = FALSE
)
}
\arguments{
\item{workbook}{the previous opened workbook}

\item{filename}{the file name to be opened}

\item{sheetname}{the sheet name}

\item{DATASheet}{the data of the new sheet}

\item{XLSX}{either XLConnect or openxls default: openxls}

\item{overwrite}{overwrite existing file - default = no}
}
\value{
returns a list with\cr
WB the Workbook\cr
the opened DataSheet\cr
}
\description{
adds a Sheet with  Dominance data to a loaded xlsx workbook
}
\examples{
{ #you can eihter use:
}
}
\author{
Knut Krueger, \email{Knut.Krueger@equine-science.de}
}
