% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.action.without.response.r
\name{change.action.without.response}
\alias{change.action.without.response}
\title{Function change.action.without.response}
\usage{
change.action.without.response(data.set, action, response, newaction, ...)
}
\arguments{
\item{data.set}{A data frame used f.e at ADI or FDI}

\item{action}{the action which is to be searched for}

\item{response}{the number of the action responing to param action}

\item{newaction}{newaction must be: data.frame("name.of.action"="test","action.number"=1,"classification"=2,"weighting"=3)\cr
or "change.only" if action is present and row kind.of.action should be changed'}

\item{\dots}{Additional parameters:
 \describe{
  \item{\bold{workbook}}{the XlConnect Workbook for the Excel file to be changed\cr
note: The workbook must be opened before}
  \item{\bold{sheet}}{the sheet name ( data will be added to be sure not to delete any data}
 }}
}
\value{
the changed data.set
}
\description{
A package to calculate Dominance Indices, print Soical Network Graphs and Music Notation Graphs.
}
\author{
Knut Krueger, \email{Knut.Krueger@equine-science.de}
}
\keyword{~ADI}
\keyword{~FDI}
