% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDominanceData.r
\name{LoadDominanceData}
\alias{LoadDominanceData}
\title{Function LoadDominanceData}
\usage{
LoadDominanceData(filename, sheetname, XLSX)
}
\arguments{
\item{filename}{the file name to be opened}

\item{sheetname}{the sheet name}

\item{XLSX}{either XLConnect or openxls or RcmdrMisc package default: openxls}
}
\value{
returns a list with\cr
WB the Workbook\cr
the opened DataSheet\cr
}
\description{
Load Dominance data with available xlsx packages and set packages
}
\examples{
{ #you can eihter use:
}
}
\author{
Knut Krueger, \email{Knut.Krueger@equine-science.de}
}
