% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.vars.R
\name{rename.vars}
\alias{rename.vars}
\title{Function rename.vars\cr
Source code from gdata as gdata is unmaintained since 2017-06-06\cr
Rename variables in a dataframe}
\usage{
rename.vars(data, from = "", to = "", info = TRUE)
}
\arguments{
\item{data}{frame to be modified.\cr}

\item{from}{character vector containing the current name of each variable to be renamed.\cr}

\item{to}{character vector containing the new name of each variable to be renamed.\cr}

\item{info}{boolean value indicating whether to print details of the removal/renaming. Defaults to TRUE.\cr}
}
\value{
returns the updated data frame with variables listed in from renamed to the corresponding element of to.\cr
}
\description{
Function rename.vars\cr
Source code from gdata as gdata is unmaintained since 2017-06-06\cr
Rename variables in a dataframe
}
\examples{
{ data <- data.frame(x=1:10,y=1:10,z=1:10)
names(data)
data <- rename.vars(data, c("x","y","z"), c("first","second","third"))
names(data)

}
}
\author{
Don MacQueen (package gdata), \email{macq\@llnl.gov.}
}
